[MOVE/#75] DTO 파일 분리
