+++ app/src/main/java/com/napzak/market/data/home/dto/HomeRecommendProductResponse.kt
@@ -10,53 +10,3 @@ data class HomeRecommendProductResponse(
     @SerialName("productSellList")
     val productSellList: List<ProductSellItem>
 )
-
-@Serializable
-data class ProductBuyItem(
-    @SerialName("productId")
-    val productId: Long,
-    @SerialName("genreName")
-    val genreName: String,
-    @SerialName("productName")
-    val productName: String,
-    @SerialName("photo")
-    val photo: String,
-    @SerialName("price")
-    val price: Int,
-    @SerialName("uploadTime")
-    val uploadTime: String,
-    @SerialName("isInterested")
-    val isInterested: Boolean,
-    @SerialName("tradeType")
-    val tradeType: String,
-    @SerialName("tradeStatus")
-    val tradeStatus: String,
-    @SerialName("isPriceNegotiable")
-    val isPriceNegotiable: Boolean,
-    @SerialName("isOwnedByCurrentUser")
-    val isOwnedByCurrentUser: Boolean
-)
-
-@Serializable
-data class ProductSellItem(
-    @SerialName("productId")
-    val productId: Long,
-    @SerialName("genreName")
-    val genreName: String,
-    @SerialName("productName")
-    val productName: String,
-    @SerialName("photo")
-    val photo: String,
-    @SerialName("price")
-    val price: Int,
-    @SerialName("uploadTime")
-    val uploadTime: String,
-    @SerialName("isInterested")
-    val isInterested: Boolean,
-    @SerialName("tradeType")
-    val tradeType: String,
-    @SerialName("tradeStatus")
-    val tradeStatus: String,
-    @SerialName("isOwnedByCurrentUser")
-    val isOwnedByCurrentUser: Boolean
-)
